refactor(pointcloud): rework point budget implementation
Since .bin and .cin have different repartitions of the points inside a node,
they must use different algorithms.
For .cin we want to draw a %-age of each node because points are evenly
distributed.
For .bin, we sort the nodes by 'importance', and stop drawing nodes when we
have enoug points.